requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
xgboost
hydra-core --upgrade
mlflow